clean ${PAI_WORK_DIR} before mv content to this folder (#3695)
* clean ${PAI_WORK_DIR} before mv content to this folder

In prod env, we saw when node down and up. The pod running on this node
may be restarted by kubelet. Even though the `restartPolicy` set to
`Never`. If that happens, the init container script will failed when
executing `mv` command. To prevent this, we clean the `${PAI_WORK_DIR}`
before mv content to this folder.

* fix review comments

* fix

* fix review comment

* add log