[Rest-server] Using hash function to generate port number (#4510)
Fix #4384
To avoid always retry due to port conflict, change to use:

(int(md5(podUid + taskPortName + portIndexForPortWithSameName)[0:12] ,16) +
 int(md5(podUid + taskPortName + portIndexForPortWithSameName)[12:24] ,16) +
 int(md5(podUid + taskPortName + portIndexForPortWithSameName)[24:32] ,16)) % 
 (globalPortEnd - globalPortStart) + globalPortStart
to generate port number.

Where globalPortEnd and globalPortStart can be configured by admin.

Notice:
This method is Best Efforts to avoid port conflict. Since we use hash function to generate port the possibility for port conflict is small.

We can not guarantee:
The ports in different jobs which running on same node will not conflict. (Same as previous)
The ports which belongs to one job but in different tasks will not conflict. (Different with previous)
All allocated ports in a single task will not conflict. (Different with previous)