remove log url prefix (#5271)
In some deploy mode. The PAI master may be behind the firewall and can not be accessed through outside. In this case, WEBPORTAL_URL will be an internal IP. Client can not use this IP to retrieve log.
Remove this prefix, let client use default domain (the gateway IP) to retrieve log,